Switches some of the interpolated fields to half precision.

+++ OSBindings/Mac/Clock Signal/ScanTarget/ScanTarget.metal
@@ -98,9 +98,9 @@ struct Line {
 struct SourceInterpolator {
 	float4 position [[position]];
 	float2 textureCoordinates;
-	float unitColourPhase;		// i.e. one unit per circle.
-	float colourPhase;			// i.e. 2*pi units per circle, just regular radians.
-	float colourAmplitude [[flat]];
+	half unitColourPhase;		// i.e. one unit per circle.
+	half colourPhase;			// i.e. 2*pi units per circle, just regular radians.
+	half colourAmplitude [[flat]];
 };
 
 struct CopyInterpolator {
